requirements/base.txt
Django==1.9.5
djangorestframework==3.6.3
psycopg2==2.7.3
django-suit==0.2.25

coreapi==2.3.1
Markdown==2.6.8
django-filter==1.0.4
django-crispy-forms==1.6.1
django-guardian==1.4.9